requirements.txt
flax
jax
matplotlib
tensorflow
numpy
absl-py
optax
pyyaml
